+++ tests/ui/test_attr_async_fn.rs
@@ -1,0 +1,9 @@
+use quickcheck_richderive::test;
+
+#[test]
+async fn t(x: u8) -> bool {
+  let _ = x;
+  true
+}
+
+fn main() {}